Boss VFX SFX Script updated
A new parent was made for the SFX VFX control in the game.
Both player and enemies inherit their respective script from it.
Enemy Script was updated to have all the necessary VFX SFX performance for the Boss
Enemies in general remain to apply this new change
Also all VFX SFX control from Combat and Health Script was passed to EnemyFX script